+++ internal/cache/filecache_test.go
@@ -1,4 +1,4 @@
-package utilities
+package cache
 
 import (
 	"fmt"
