+++ internal/templates/types.go
@@ -1,26 +1,15 @@
 package templates
 
+import "github.com/erindatkinson/slack-emojinator/internal/cache"
+
 type Docs struct {
 	Namespace string
-	Pages     []*EmojiPage
+	Pages     []*cache.EmojiPage
 }
 
 type ReleaseData struct {
 	Start string
 	End   string
-}
-
-type EmojiPage struct {
-	Name     string
-	Count    int
-	NextPage string
-	PrevPage string
-	Emojis   []EmojiItem
-}
-
-type EmojiItem struct {
-	Name    string
-	DocPath string
 }
 
 type RanksData struct {
